refactor(App): TS 마이그레이션
